backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
google-generativeai==0.3.2
requests==2.31.0
Werkzeug==2.3.7
scikit-learn
pandas
numpy
scipy
